Add config option for include_RO and delete redundant dye desal flowsheet
