+++ watertap/examples/flowsheets/case_studies/wastewater_resource_recovery/dye_desalination/tests/test_dye_desalination.py
@@ -21,26 +21,27 @@ from watertap.examples.flowsheets.case_studies.wastewater_resource_recovery.dye_
     build,
     set_operating_conditions,
     initialize_system,
+    initialize_costing,
     solve,
     add_costing,
-    display_costing,
+    optimize_operation,
     display_results,
+    display_costing,
 )
-
 
 solver = get_solver()
 
 
-class TestDyeFlowsheet:
+class TestDyewithROFlowsheetwithPretreatment:
     @pytest.fixture(scope="class")
     def system_frame(self):
-        m = build(include_gac=False)
+        m = build(include_pretreatment=True)
         return m
 
     @pytest.mark.unit
     def test_build(self, system_frame):
         m = system_frame
-        assert_degrees_of_freedom(m, 11)
+        assert_degrees_of_freedom(m, 28)
         assert_units_consistent(m)
 
     @pytest.mark.component
@@ -49,41 +50,66 @@ class TestDyeFlowsheet:
         set_operating_conditions(m)
         initialize_system(m)
 
-        # test feed conditions
+        # test feed
         assert pytest.approx(77.607, rel=1e-3) == value(
             m.fs.feed.flow_mass_comp[0, "H2O"]
         )
 
-        assert pytest.approx(0.015556, rel=1e-3) == value(
-            m.fs.feed.flow_mass_comp[0, "dye"]
-        )
+        assert pytest.approx(2, rel=1e-5) == value(m.fs.feed.conc_mass_comp[0, "tds"])
 
-        assert pytest.approx(0.155556, rel=1e-3) == value(
-            m.fs.feed.flow_mass_comp[0, "tds"]
-        )
+        assert pytest.approx(0.2, rel=1e-5) == value(m.fs.feed.conc_mass_comp[0, "dye"])
 
-        # test nanofiltration block
-        assert pytest.approx(316.219, rel=1e-3) == value(
-            m.fs.dye_separation.nanofiltration.area
+        # test wwtp
+        assert pytest.approx(0.14, rel=1e-5) == value(
+            m.fs.pretreatment.wwtp.energy_electric_flow_vol_inlet
         )
 
         # test pump block
-        assert pytest.approx(7, rel=1e-6) == value(
+        assert pytest.approx(7, rel=1e-5) == value(
             m.fs.dye_separation.P1.applied_pressure[0]
+        )
+
+        # test nanofiltration
+        assert pytest.approx(315.9843, rel=1e-5) == value(
+            m.fs.dye_separation.nanofiltration.area
+        )
+
+        # check products
+        assert pytest.approx(0, rel=1e-6) == value(
+            m.fs.wwt_retentate.flow_mass_comp[0, "dye"]
+        )
+        assert pytest.approx(0.026809, rel=1e-3) == value(
+            m.fs.concentrated_dye.flow_mass_comp[0, "tds"]
+        )
+
+        assert pytest.approx(0.965, rel=1e-5) == value(
+            m.fs.permeate.flow_mass_phase_comp[0, "Liq", "H2O"]
+        )
+
+        assert pytest.approx(0.035, rel=1e-5) == value(
+            m.fs.brine.flow_mass_phase_comp[0, "Liq", "TDS"]
         )
 
     @pytest.mark.component
     def test_solve(self, system_frame):
         m = system_frame
+
         results = solve(m)
 
-        # check each product
-        assert pytest.approx(64.4135, rel=1e-6) == value(
-            m.fs.permeate.flow_mass_comp[0, "H2O"]
+        # check products
+        assert pytest.approx(0.02894, rel=1e-3) == value(
+            m.fs.wwt_retentate.flow_mass_comp[0, "tds"]
+        )
+        assert pytest.approx(13.1931, rel=1e-3) == value(
+            m.fs.concentrated_dye.flow_mass_comp[0, "H2O"]
         )
 
-        assert pytest.approx(0.015426, rel=1e-3) == value(
-            m.fs.dye_retentate.flow_mass_comp[0, "dye"]
+        assert pytest.approx(32.2185, rel=1e-5) == value(
+            m.fs.permeate.flow_mass_phase_comp[0, "Liq", "H2O"]
+        )
+
+        assert pytest.approx(32.1950, rel=1e-5) == value(
+            m.fs.brine.flow_mass_phase_comp[0, "Liq", "H2O"]
         )
 
     @pytest.mark.component
@@ -91,11 +117,15 @@ class TestDyeFlowsheet:
         m = system_frame
 
         add_costing(m)
+        initialize_costing(m)
+        optimize_operation(m)
+
         results = solve(m)
         assert_optimal_termination(results)
 
-        # check values
-        assert pytest.approx(19.626110, rel=1e-3) == value(m.fs.LCOT)
+        # check costing
+        assert pytest.approx(27.821289, rel=1e-3) == value(m.fs.LCOW)
+        assert pytest.approx(19.835558, rel=1e-3) == value(m.fs.LCOT)
 
     @pytest.mark.component
     def test_display(self, system_frame):
@@ -104,7 +134,200 @@ class TestDyeFlowsheet:
         display_costing(m)
 
 
-class TestDyeFlowsheet_withGAC:
+class TestDyewithROFlowsheetDefault:
+    @pytest.fixture(scope="class")
+    def system_frame(self):
+        m = build()
+        return m
+
+    @pytest.mark.unit
+    def test_build(self, system_frame):
+        m = system_frame
+        assert_degrees_of_freedom(m, 24)
+        assert_units_consistent(m)
+
+    @pytest.mark.component
+    def test_set_operating_conditions(self, system_frame):
+        m = system_frame
+        set_operating_conditions(m)
+        initialize_system(m)
+
+        # test feed
+        assert pytest.approx(77.607, rel=1e-3) == value(
+            m.fs.feed.flow_mass_comp[0, "H2O"]
+        )
+
+        assert pytest.approx(2, rel=1e-5) == value(m.fs.feed.conc_mass_comp[0, "tds"])
+
+        assert pytest.approx(0.2, rel=1e-5) == value(m.fs.feed.conc_mass_comp[0, "dye"])
+
+        # test pump block
+        assert pytest.approx(7, rel=1e-5) == value(
+            m.fs.dye_separation.P1.applied_pressure[0]
+        )
+
+        # test nanofiltration
+        assert pytest.approx(316.096, rel=1e-5) == value(
+            m.fs.dye_separation.nanofiltration.area
+        )
+
+        # check products
+        assert pytest.approx(0.032937, rel=1e-3) == value(
+            m.fs.concentrated_dye.flow_mass_comp[0, "tds"]
+        )
+
+        assert pytest.approx(0.965, rel=1e-5) == value(
+            m.fs.permeate.flow_mass_phase_comp[0, "Liq", "H2O"]
+        )
+
+        assert pytest.approx(0.035, rel=1e-5) == value(
+            m.fs.brine.flow_mass_phase_comp[0, "Liq", "TDS"]
+        )
+
+    @pytest.mark.component
+    def test_solve(self, system_frame):
+        m = system_frame
+
+        results = solve(m)
+
+        # check products
+        assert pytest.approx(13.1931, rel=1e-3) == value(
+            m.fs.concentrated_dye.flow_mass_comp[0, "H2O"]
+        )
+
+        assert pytest.approx(32.2212, rel=1e-5) == value(
+            m.fs.permeate.flow_mass_phase_comp[0, "Liq", "H2O"]
+        )
+
+        assert pytest.approx(32.1923, rel=1e-5) == value(
+            m.fs.brine.flow_mass_phase_comp[0, "Liq", "H2O"]
+        )
+
+    def test_costing(self, system_frame):
+        m = system_frame
+
+        add_costing(m)
+        initialize_costing(m)
+        optimize_operation(m)
+
+        results = solve(m)
+        assert_optimal_termination(results)
+
+        # check costing
+        assert pytest.approx(27.719244, rel=1e-3) == value(m.fs.LCOW)
+        assert pytest.approx(19.774679, rel=1e-3) == value(m.fs.LCOT)
+
+    @pytest.mark.component
+    def test_display(self, system_frame):
+        m = system_frame
+        display_results(m)
+        display_costing(m)
+
+
+class TestDyewithROFlowsheetwithDewatering:
+    @pytest.fixture(scope="class")
+    def system_frame(self):
+        m = build(include_dewatering=True)
+        return m
+
+    @pytest.mark.unit
+    def test_build(self, system_frame):
+        m = system_frame
+        assert_degrees_of_freedom(m, 27)
+        assert_units_consistent(m)
+
+    @pytest.mark.component
+    def test_set_operating_conditions(self, system_frame):
+        m = system_frame
+        set_operating_conditions(m)
+        initialize_system(m)
+
+        # test feed
+        assert pytest.approx(77.607, rel=1e-3) == value(
+            m.fs.feed.flow_mass_comp[0, "H2O"]
+        )
+
+        assert pytest.approx(2, rel=1e-5) == value(m.fs.feed.conc_mass_comp[0, "tds"])
+
+        assert pytest.approx(0.2, rel=1e-5) == value(m.fs.feed.conc_mass_comp[0, "dye"])
+
+        # test dewaterer block
+        assert pytest.approx(0.99, rel=1e-5) == value(
+            m.fs.dewaterer.split_fraction[0, "precipitant", "dye"]
+        )
+
+        # test pump block
+        assert pytest.approx(7, rel=1e-5) == value(
+            m.fs.dye_separation.P1.applied_pressure[0]
+        )
+
+        # test nanofiltration
+        assert pytest.approx(316.0961, rel=1e-5) == value(
+            m.fs.dye_separation.nanofiltration.area
+        )
+
+        # check products
+        assert pytest.approx(1, rel=1e-3) == value(
+            m.fs.concentrated_dye.flow_mass_comp[0, "tds"]
+        )
+
+        assert pytest.approx(0.965, rel=1e-5) == value(
+            m.fs.permeate.flow_mass_phase_comp[0, "Liq", "H2O"]
+        )
+
+        assert pytest.approx(0.035, rel=1e-5) == value(
+            m.fs.brine.flow_mass_phase_comp[0, "Liq", "TDS"]
+        )
+
+    @pytest.mark.component
+    @pytest.mark.requires_idaes_solver
+    def test_solve(self, system_frame):
+        m = system_frame
+
+        results = solve(m)
+
+        # check products
+        assert pytest.approx(13.0612, rel=1e-3) == value(
+            m.fs.treated.flow_mass_comp[0, "H2O"]
+        )
+
+        assert pytest.approx(0.131931, rel=1e-3) == value(
+            m.fs.concentrated_dye.flow_mass_comp[0, "H2O"]
+        )
+
+        assert pytest.approx(32.2212, rel=1e-5) == value(
+            m.fs.permeate.flow_mass_phase_comp[0, "Liq", "H2O"]
+        )
+
+        assert pytest.approx(32.1923, rel=1e-5) == value(
+            m.fs.brine.flow_mass_phase_comp[0, "Liq", "H2O"]
+        )
+
+    @pytest.mark.component
+    @pytest.mark.requires_idaes_solver
+    def test_costing(self, system_frame):
+        m = system_frame
+
+        add_costing(m)
+        initialize_costing(m)
+        optimize_operation(m)
+
+        results = solve(m)
+        assert_optimal_termination(results)
+
+        # check costing
+        assert pytest.approx(1.08502, rel=1e-3) == value(m.fs.LCOW)
+        assert pytest.approx(-0.151027, rel=1e-3) == value(m.fs.LCOT)
+
+    @pytest.mark.component
+    @pytest.mark.requires_idaes_solver
+    def test_display(self, system_frame):
+        m = system_frame
+        display_results(m)
+        display_costing(m)
+
+
+class TestDyewithROFlowsheetwithGAC:
     @pytest.fixture(scope="class")
     def system_frame(self):
         m = build(include_gac=True)
@@ -113,7 +336,7 @@ class TestDyeFlowsheet_withGAC:
     @pytest.mark.unit
     def test_build(self, system_frame):
         m = system_frame
-        assert_degrees_of_freedom(m, 30)
+        assert_degrees_of_freedom(m, 43)
 
     @pytest.mark.component
     def test_set_operating_conditions(self, system_frame):
@@ -122,44 +345,42 @@ class TestDyeFlowsheet_withGAC:
         assert_units_consistent(m)
         initialize_system(m)
 
-        # test feed conditions
+        # test feed
         assert pytest.approx(77.607, rel=1e-3) == value(
             m.fs.feed.flow_mass_comp[0, "H2O"]
         )
 
-        assert pytest.approx(0.015556, rel=1e-3) == value(
-            m.fs.feed.flow_mass_comp[0, "dye"]
-        )
+        assert pytest.approx(2, rel=1e-5) == value(m.fs.feed.conc_mass_comp[0, "tds"])
 
-        assert pytest.approx(0.155556, rel=1e-3) == value(
-            m.fs.feed.flow_mass_comp[0, "tds"]
-        )
-
-        # test nanofiltration block
-        assert pytest.approx(316.219, rel=1e-3) == value(
-            m.fs.dye_separation.nanofiltration.area
-        )
+        assert pytest.approx(0.2, rel=1e-5) == value(m.fs.feed.conc_mass_comp[0, "dye"])
 
         # test pump block
-        assert pytest.approx(7, rel=1e-6) == value(
+        assert pytest.approx(7, rel=1e-5) == value(
             m.fs.dye_separation.P1.applied_pressure[0]
         )
 
+        # test nanofiltration
+        assert pytest.approx(316.0961, rel=1e-5) == value(
+            m.fs.dye_separation.nanofiltration.area
+        )
+
     @pytest.mark.component
+    @pytest.mark.requires_idaes_solver
     def test_solve(self, system_frame):
         m = system_frame
+
         results = solve(m)
 
-        # check each product
-        assert pytest.approx(64.4135, rel=1e-6) == value(
-            m.fs.permeate.flow_mass_comp[0, "H2O"]
+        # check products
+        assert pytest.approx(32.22122, rel=1e-6) == value(
+            m.fs.permeate.flow_mass_phase_comp[0, "Liq", "H2O"]
         )
 
         assert pytest.approx(13.19313, rel=1e-3) == value(
             m.fs.treated.flow_mass_phase_comp[0, "Liq", "H2O"]
         )
 
-        assert pytest.approx(0.00092002, rel=1e-3) == value(
+        assert pytest.approx(0.000485234, rel=1e-3) == value(
             m.fs.treated.flow_mass_phase_comp[0, "Liq", "dye"]
         )
 
@@ -168,29 +389,35 @@ class TestDyeFlowsheet_withGAC:
         )
 
         assert pytest.approx(0, rel=1e-3) == value(
-            m.fs.adsorbed_dye.flow_mass_phase_comp[0, "Liq", "H2O"]
+            m.fs.concentrated_dye.flow_mass_phase_comp[0, "Liq", "H2O"]
         )
 
-        assert pytest.approx(0.014506, rel=1e-3) == value(
-            m.fs.adsorbed_dye.flow_mass_phase_comp[0, "Liq", "dye"]
+        assert pytest.approx(0.0149412, rel=1e-3) == value(
+            m.fs.concentrated_dye.flow_mass_phase_comp[0, "Liq", "dye"]
         )
 
         assert pytest.approx(0, rel=1e-3) == value(
-            m.fs.adsorbed_dye.flow_mass_phase_comp[0, "Liq", "tds"]
+            m.fs.concentrated_dye.flow_mass_phase_comp[0, "Liq", "tds"]
         )
 
     @pytest.mark.component
+    @pytest.mark.requires_idaes_solver
     def test_costing(self, system_frame):
         m = system_frame
 
         add_costing(m)
+        initialize_costing(m)
+        optimize_operation(m)
+
         results = solve(m)
         assert_optimal_termination(results)
 
-        # check values
-        assert pytest.approx(-0.781, rel=1e-3) == value(m.fs.LCOT)
+        # check costing
+        assert pytest.approx(0.415656, rel=1e-3) == value(m.fs.LCOW)
+        assert pytest.approx(-0.765208, rel=1e-3) == value(m.fs.LCOT)
 
     @pytest.mark.component
+    @pytest.mark.requires_idaes_solver
     def test_display(self, system_frame):
         m = system_frame
         display_results(m)
