+++ watertap/examples/flowsheets/case_studies/wastewater_resource_recovery/dye_desalination/tests/test_dye_sweep.py
@@ -28,14 +28,12 @@ def test_dye_sweep(case_num, tmp_path):
     cwd = os.getcwd()
     os.chdir(tmp_path)
     # test every other sweep with RO
-    withRO = bool(case_num == 1)
     temp = tempfile.NamedTemporaryFile(delete=False)
     temp.close()
     results, params, model = run_analysis(
         case_num,
         nx=1,
         interpolate_nan_outputs=False,
-        withRO=withRO,
         output_filename=temp.name,
     )
     os.remove(temp.name)
@@ -49,16 +47,8 @@ def test_dye_sweep2(case_num, tmp_path):
     cwd = os.getcwd()
     os.chdir(tmp_path)
     # test only with RO
-    run_analysis(case_num, nx=1, interpolate_nan_outputs=False, withRO=True)
+    run_analysis(case_num, nx=1, interpolate_nan_outputs=False)
     os.chdir(cwd)
-    return
-
-
-@pytest.mark.integration
-def test_withRO_condition():
-    # test a case without RO for a sweep that requires RO
-    with pytest.raises(Exception):
-        run_analysis(15, 1, interpolate_nan_outputs=False, withRO=False)
     return
 
 
@@ -66,5 +56,5 @@ def test_withRO_condition():
 def test_out_of_range_cases():
     # test a case number that is not included in the dye sweep file
     with pytest.raises(Exception):
-        run_analysis(30, 1, interpolate_nan_outputs=False, withRO=False)
+        run_analysis(30, 1, interpolate_nan_outputs=False)
     return
